transcript_minutes_api2/pyproject.toml
[tool.poetry]
name = "transcript-minutes-api2"
version = "0.1.0"
description = "Transcript to Meeting Minutes API with JWT Authentication and Microsoft Graph Integration"
authors = ["Devin AI <devin-ai-integration[bot]@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
openai = "^1.3.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
httpx = "^0.25.2"
msgraph-core = "^1.0.0"
azure-identity = "^1.15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
